+++ backend/src/main/java/com/devtalk/model/MessageType.java
@@ -1,4 +1,4 @@
-package com.devtalk.backend.model;
+package com.devtalk.model;
 
 public enum MessageType {
     CHAT, JOIN, LEAVE
