+++ backend/src/test/java/com/devtalk/BackendApplicationTests.java
@@ -1,4 +1,4 @@
-package com.devtalk.backend;
+package com.devtalk;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
